fix(powl): drop redundant u64 cast in scheduler_tick_guarded
cargo clippy -p bcinr-powl flagged clippy::unnecessary_cast on the new
scheduler_tick_guarded's sat_bit computation (pred_satisfied already
returns u64, so `as u64` was a no-op). Fixed; re-ran cargo test -p
bcinr-powl --lib: 209 passed, 0 failed.

+++ crates/bcinr-powl/src/scheduler.rs
@@ -374,7 +374,7 @@ pub fn scheduler_tick_guarded<S: ConcurrencySelector>(
         let effective_pred = (join_effective & is_join) | (op.pred_mask & !is_join);
 
         let sat = pred_satisfied(new_done, effective_pred);
-        let sat_bit = (sat & 1) as u64;
+        let sat_bit = sat & 1;
         let selected_bit = (selected_mask >> i) & 1;
         let fire_mask = u64::wrapping_sub(0, sat_bit & selected_bit) & bit;
